Control automatic opening, change its default to false
Previously, notifications received were displayed as heads up if the
notification is in foreground AND the notification has no in-app.

To improve usage of deep links and simplify what happens when a push
notification is received, we now systematically show the notification
as a heads up if the application is in foreground.
For flexibility, an automatic open feature has been added too.