Notification update using tag
Controls whether a notification should update an existing one, or be
displayed as a new one.

If `_wp.alert.tag` is present, it governs notification updates, with
null meaning a new notification is always displayed.
If absent, then we use the campaignId in `_wp.c` and apply the same
rule.
This permits to prevent collapsing inside a single campaign.