Use manifest merger to ease integration
I had to modify a gradle task so that AAPT accepts the manifest
reference to a resource, so I introduced a generic notification icon
that can be reused as is by apps.
The icon comes from Google Material Ssytem icons pack.

Appart from referencing the SDK as a dependency, the developer now needs
to add the following to its build.gradle:

    android {
        defaultConfig {
            manifestPlaceholders = [
                    wonderpushNotificationIcon: '@drawable/notification_icon',
                    wonderpushDefaultActivity: '.MainActivity'
            ]
        }
    }

`wonderpushNotificationIcon` can be set to
`@drawable/ic_notifications_white_24dp` to use the built-in bell icon.
But `@drawable/ic_launcher` or a specific white-on-transparent icon is
best.

No modifications of the AndroidManifest.xml is needed, except if for the
geolocation permission that is left as a choice to the developer.