Handle actions even if targetUrl does not open the app
The `@NOTIFICATION_OPENED` event was not sent if the `targetUrl` asked
`WonderPushService` to open the browser instead.

Reworked the service/activity notification handling to only handle the
in-app in the activity.

Also added the `wonderpush://notificationOpen/noop` url to prevent the
application from showing and any in-app to trigger. Useful for "Cancel"
notification action buttons, for instance.