Fix spurious @APP_OPEN on background SDK interactions
Calling `putInstallationCustomProperties()` or `trackEvent()` in the
Application class caused `@APP_OPEN` events to be created although the
application was never opened (only a notification was potentially
shown).

On pre-ICS (API 14) devices, we have no easy was to tell the difference,
so the bug is still here. But as of 2017-04-03 this accounts for 0.9% of
all Android devices.