Do not enforce default channel creation to avoid later creation by the application
In the previous code, at the very first launch of the application under
Android O, we would create the default channel inside Android,
preventing the application to create configure it beforehand.

We now only enforce its existence only when using it (once Android O
APIs will be published).

This also permits setting the default channel id, and then let the
application create it itself in Android.