Do not force notification title from Android 5 onward
Since Android 5 the application identity is always displayed using the
small icon and explicit when long-pressing the notification.
Since Android 7 the application name is even explicitly displayed.
This makes forcing the notification title useless from Android 5 onward.
We keep it on Android 4 (~10% of official Android market share) for
UI clarity.