Fix launching of default activity by using PackageManager.getLaunchIntentForPackage()
The issue with Intent.resolveActivity() with an implicit intent is that
it uses MATCH_DEFAULT_ONLY and the launch activity are not required to
have the DEFAULT category, so it matched nothing, logged an error and
did not start the application.

PackageManager.getLaunchIntentForPackage() however does not use
MATCH_DEFAULT_ONLY and finds an activity and returns an explicit intent
on which Intent.resolveActivity() returns the component set.