Fix removeProperty type mismatches
A deserialized JSONObject/JSONArray and a manually constructed one can
differ in the implementation of Number used to store numeric values.
This caused removeProperty(1L) to never work because the stored
properties are then deserialized using an Integer instead of a Long.

Enforcing an extra serialization/deserialization is hence the best way
to enforce that the types and values will match with what's stored.