Update picasso from v2.71828 to v2.8 for AndroidX
Without this change for projects that to not set
`android.enableJetifier=true` in their `gradle.properties` file, they
would end up with both AndroidX and Legacy Support libraries, possibly
resulting in duplicate classes I believe.

Note that the fancy versioning of Picasso may impose that 2.71828 is
ulterior to 2.8 using semantic versioning (or any reasonable versioning,
for that matter).