Fix for Android 12 (API 31) Notification trampoline restrictions
Bug behavior:
-------------

Applications targeting Android SDK API 31 that do not update to this
SDK version will no longer see any activity opened upon clicking on a
notification.

See https://developer.android.com/about/versions/12/behavior-changes-12#notification-trampolines.

Corresponding warning in logcat:

    Indirect notification activity start (trampoline) from PACKAGE_NAME, this should be avoided for performance reasons.

Explanation and fix:
--------------------

We previously used a Service to handle push notification clicks and
redirect to the proper screen. As this is now restricted for
applications targeting Android SDK API 31, we changed the time when the
deeplink is resolved from the notification click to the notification
receipt, and add our tracking activity on top of it to continue to
properly track clicked notifications.

Note if you implement `WonderPushDelegate`:
-------------------------------------------

The `urlForDeepLink()` method is now called upon _receiving_ the
notification, not _opening_ the relevant deeplink upon _clicking_ it.

This means that if you were used to return `null` to indicate that you
took care of starting the proper activity, you are at risk of opening
the deeplink when the notification is received instead of clicked.

Returning `null` keeps the meaning of telling WonderPush to not open any
activity.