Track a new @APP_OPEN even when an activity is resumed after a long time
When an activity is resumed, even if it was not started moments before,
we should still count this as a new session.

Related: https://developer.android.com/about/versions/12/behavior-changes-all#back-press.

Android 12 no longer stops the root activity when pressing back to align
the behavior with putting the application in background.

This makes that behavior all the more frequent.