Support delaying on launch/foreground in-apps on identified custom splash or routing activities
Add the following metadata to your `<activity>` in your
`AndroidManifest.xml` to have the "App launch" and "On foreground"
in-app triggers be delayed until landing on a normal activity.

    <meta-data android:name="com.wonderpush.sdk.iam.ignoreForeground" android:value="true" />

Mark your custom splash activity with this metadata. You can mark
routing or dialog activities too. Any activity that is typically shown
for a short time at the start of your application is a good candidate.
This avoids displaying an in-app message on the next app launch or
foreground event only to see it dismissed when your application
automatically navigates to another activity.