ActivityLifecycleMonitor notes system time before adjusting it once async
Reading synced time does IO wit SharedPreferences, which has a high
penalty for the very first read, so it's better to do this IO
asynchronously. However to avoid reading the current time only once the
async process has started we split the measuring of current time from
the adjustment to make it synced with the server.