Fix deadlock with WonderPushUserPreferences initialization
Because the call to `WonderPushUserPreferences.initialize()` is now
called asynchronously, it is possible that the developer's
`Application.onCreate()` (or other method) calls a
`WonderPushUserPreferences` method before the class is initialized.

The issue is that because all methods are `synchronized`, when
`WonderPushUserPreferences.initialize()` is called, or any other
`static` methods called further down its execution are called, they wait
for the other `public synchronized` class methods from returning,
causing a deadlock.

`private` methods are no longer `synchronized`, so is the `initialize()`
method. This way it can be called after other `public synchronized`
methods and it can unblock them, and it can call any `private` method at
will to perform it's duty.