Increase ScheduledExecutor core pool size from 1 to 4
If a `Runnable` waits on a lock, no other task would run previously, now
we have a bit more chance to have another task unlocking the blocked
one.

For instance if the order of the deferred calls in
`WonderPush.initialize(Context, String, String)` was different, like if
something reads from `JSONSyncInstallation` before it is initialized,
that call would synchronously wait for the latter to initialize, but the
its deferred initialization would never kick in.

So as a precaution measure, I am increasing the number of tasks running
in parallel to avoid deadlocks.
`safeDefer` never was a means to ensure serial access, and we protect
many code paths in synchronized sections already, so this should not
create uncontrolled concurrency issues.

Sadly the `ScheduledThreadPoolExecutor` behind
`Executors.newScheduledThreadPool()` does not reuse existing worker
threads until the core pool size has been reached, so I am not using a
higher value.