Call SubscriptionStatusListener everytime we refresh the subscription status, not just when it changes
The goal is to make sure the `safeDeferWithSubscription` tasks don't get
stuck if we determined to be opt-in even if seem to already knew it.

I think there can be a code path where we can use some userId and be
opt-out (due to `unsubscribeFromNotifications()`), hence queue tasks,
then change userId and refresh the subscription but
`refreshSubscriptionStatus()` starts reading that that other profile is
already opt-in (due to `subscribeToNotifications()`), hence no
subscription status change is observed.

This could lead to some events not being tracked.