Implement initializeAndRememberCredentials(), enabling complex init scenarios, especially with framework integrations
`initializeAndRememberCredentials()` not only initializes the WonderPush
SDK but also stores the given credentials so that next time the SDK must
initialize, it can automatically do so on its own without relying on
being provided a value through BuildConfig fields, resources, manifest
metadata or custom code.

In complex customer scenarios, the credentials to be used cannot be
statically determined and are not available at startup, typically not
until the end-user selects a country or brand.

The WonderPush SDK however needs to be initialized in a few scenarios,
for instance when a push notification is received. When the SDK is
integrated through the use of some framework, it is not always easy or
feasible to have a synchronous communication with the code within the
framework at such time, so the SDK is now able to automatically reuse
remembered credentials.

Because the credentials are necessarily stored in a format readily
accessible for the application (although not to other applications),
it is necessary to opt into this behaviour by:

1. Disable automatic initialization by removing the Client ID and
   Client Secret from the BuildConfig fields, resources or manifest
   metadata. You can alternatively use the value `USE_REMEMBERED`.
2. Call `initializeAndRememberCredentials()` with the Client ID and
   Client Secret.

Providing `null` values to `initializeAndRememberCredentials()` will
remove the previously stored values.