Support direct calls to `WonderPush.initialize(Context, "USE_REMEMBERED", "USE_REMEMBERED")`
`USE_REMEMBERED` was only supported in the automatic initialization, but
this may not be how framework integrations that avoid native code will
be using it.

It previously tried to use that value as is for Client ID, and failed to
resolve the remote configuration.

This commit aligns the behavior with the iOS SDK behavior.